dev and prod profiles, exception handling

+++ product_service/src/main/java/org/msia_projekt/product_service/services/ArticleService.java
@@ -1,12 +1,12 @@
 package org.msia_projekt.product_service.services;
 
+import org.msia_projekt.product_service.ArticleDoesntExistsException;
 import org.msia_projekt.product_service.entities.Article;
 import org.msia_projekt.product_service.repositories.IArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ArticleService {
@@ -19,23 +19,30 @@ public class ArticleService {
     }
 
     public Article createArticle(Article article) {
-        return this.iArticleRepository.saveAndFlush(article);
+        return this.iArticleRepository.save(article);
     }
 
-    public Article updateArticle(Long id, Article article) {
-        article.setId(id);
-        return this.iArticleRepository.saveAndFlush(article);
+    public Article updateArticle(Long id, Article updatedArticle) {
+        if (!this.iArticleRepository.existsById(id))
+            throw new ArticleDoesntExistsException(id);
+
+        updatedArticle.setId(id);
+        return this.iArticleRepository.save(updatedArticle);
     }
 
     public void deleteArticleById(Long id) {
+        if (!this.iArticleRepository.existsById(id))
+            throw new ArticleDoesntExistsException(id);
+
         this.iArticleRepository.deleteById(id);
     }
 
-    public Optional<Article> readArticleById(Long id) {
-        return this.iArticleRepository.findById(id);
+    public Article readArticleById(Long id) {
+        return this.iArticleRepository.findById(id).orElseThrow(() -> new ArticleDoesntExistsException(id));
     }
 
     public List<Article> readAllArticles() {
         return this.iArticleRepository.findAll();
     }
+
 }
